starboard/types.h: Remove SSIZE_T provisions/fallbacks (#4382)
A cursory search shows that SSIZE_T is defined in all //starboard
platforms. This CL aims to remove it, but before that I want to
see whether any bot configuration is covering this path.

SSIZE_MAX is certainly not used as per code search so can be
removed right away.

b/366064934