Add play-services-ads-identifier-18.2.0 to the dependency and re-enable AdvertisingId.java. (#4579)
We submit
https://chromium-review.googlesource.com/c/chromium/src/+/6086302 to
chromium repo, then because of privacy concern,
[go/no-ads-identifier](http://goto.google.com/no-ads-identifier), we
reverted it with
https://chromium-review.googlesource.com/c/chromium/src/+/6092014.

Even though it is reverted, the aar binary is uploaded to CIPD. So
instead of submit the aar binary file directly to the cobalt repo, we
reuse the CIPD infrastructure to pull the library.

The files are autogenerated by using
`./third_party/android_deps/fetch_all.py --local` command on Chromium
repo 's main branch, then copied cover to the cobalt repo with a minor
change in the DEPS file.

b/377049113

Co-authored-by: Colin Liang <colinliang@google.com>