Starboard: support --v for verbosity (#4583)
Starboard builds and in particular the nplb integration test support a command line flag
--min_log_level=(info|warning|error|fatal). Chromium binaries prefer in general --v
(or -v) for this purpose.

This CL adds logic to handle --v; to preserve previous behaviour, if and when both
--min_log_level and --v are present, the former takes precedence.

This CL also add unit tests for the translation between flag arguments and log level.
First instance of starboard_unittests !

Bug: 378879686
Bug: 384819454