[media] Support media suspend and resume (#4580)
CobaltActivity adopts ContentShellActivity, but ContentShell.apk didn't
set WebContents correctly when the activity onStop(), resulting in it
continued to play the video after pushing to the background.

Set WebContents to onHide() when the activity onStop(). This allows
Cobalt.apk to have the same behavior as Chrome when suspend/resume the
app on android.

Chrome upstream code is:
https://chromium-review.googlesource.com/c/chromium/src/+/6092099

b/384086780
b/383153069
b/326497953