[media] Move concrete SbAudioSink impl to namespace (#4634)
SbAudioSink is a typedef to SbAudioSinkPrivate*, and the concrete
implementation of SbAudioSinkPrivate was in the global namespace. Now
it's moved into a new class named SbAudioSinkImpl in the namespace
starboard::shared::starboard::audio_sink, and it's derived from
SbAudioSinkPrivate, which is now an interface.

This allows to have multiple SbAudioSink implementations in different
namespaces.

b/327287075