Refactor CrashAnnotator impl as a Cobalt service (#4709)
This is in response to feedback from mcasas@, and others: since the
CrashAnnotator is not a core service, it shouldn't be implemented under
//content. //cobalt/services seems like a good place since the
CrashAnnotator service is for use by Cobalt in particular.

The use of DocumentService was recommended by dcheng@ and ortuno@ as a
way to scope our Mojo impl's lifetime to the document / RenderFrameHost.

b/383301493