cobalt/browser: trivially extend ShellBrowserMainParts logic (#4808)
This CL overrides ShellContentBrowserClient::CreateBrowserMainParts()
to create a local, trivial CobaltBrowserMainParts.

This CL introduces no functional changes, its aim is to prepare the
groundwork for adding UMA reporting logic (*) in Chrobalt.

Aside from that, a few TODO()s are added to remember that long-term
Cobalt should not extend Shell(Content|Browser) classes, but directly
the (Content|Browser) ones.  Finally some constants are moved closer
to their callsites and the empty ctor/dtor are replaced with `=
default`.

(*) UMA collection is already active in the system, as can be seen
navigating to chrome:histograms;

Bug: b/372559349
Bug: b/390021478