cobalt/browser/metrics: Skeleton CobaltMetricsServiceClient (#4934)
This CL creates a skeleton CobaltMetricsServiceClient, a class
that acts as the embedder's counterpart to MetricsService, the
main control class for e.g. UMA reports. CobaltMetricsServiceClient
instructs the latter to e.g. start recording and uploading UMA
reports at a configurable rate (GetStandardUploadInterval(), 5
minutes).

CobaltMetricsServiceClient (CSMSC) is heavily inspired on the
simple CastMetricsServiceClient (other embedders' logic is much
more involved).

Some gotchas/details worth mentioning:

- CMSC Relies on an empty CobaltMetricsLogUploader; this class
should be implemented to not upload the metrics blob to an URL,
but to send it "up" to the WebApp (Kabuki).
- PrefRegistry/Service play an important role in embedders; here a
PrefRegistrySimple/InMemoryPrefStore pair are used. These are
unlikely to be enough in the long run.
- Consent and Reporting enabled are stubbed to always yes. This
will probably need revisiting :)
- All BUILD-dependencies are in a new source_set(); this is to
identify dependencies introduced specifically by this new class,
facilitating in time the removal of the dependency on
content_shell_lib

Bug: b/372559349