Get Runtime Permission for AUDIO_RECORD (#5055)
b/383379319

Get Runtime permissions for AUDIO_RECORD.

Although we have "android.permission.RECORD_AUDIO" in the
AndroidManifest, we must request permissions at runtime.

This PR reintroduces the PermissionRequestor. It will be triggered by a
call to CheckMediaAccessPermission() in Blink code, whereupon a JNI call
is issued which goes through StarboardBridge to the PermissionRequestor.
It will check if PERMISSION_GRANTED for RECORD_AUDIO. If its not
granted, it will run ActivityCompat.requestPermissions() to ask the
user, and then handle the necessary onRequestPermissionsResult() if
granted. This implements the RequestPermission flow as described here:
https://developer.android.com/training/permissions/requesting#java

In previous Cobalt, the PermissionRequester had to call back to the
SbMicrophone implementation to signal recording to start. That is not
longer necessary and has been cleaned up.

Currently this modifies code in content/shell/browser/shell.* -
b/401397483 tracks the work to add our own CobaltWebContentsDelegate
where these modifications can move to.

Tested with
https://storage.googleapis.com/cobalt-tools.appspot.com/dpadtest.html
Testing instructions. Uninstall Cobalt to ensure no permissions are left
over. Do a fresh install, then try to use mic on youtube, you should be
asked for permission. Once allowed, you should be able to use [the test
page to use
mic.](https://storage.googleapis.com/cobalt-tools.appspot.com/dpadtest.html)