Bypass Chromium A11Y and add Cobalt's A11Y (#5180)
This PR hacks the browser to bypass Chromium's A11Y navigation engine,
and adds Cobalt's A11Y navigation helper.

Add CobaltA11yHelper to be a member of CobaltActivity, because
CobaltActivity should manage the lifecycle of CobaltA11yHelper. But we
also need to pass in a weak reference of CobaltActivity to
CobaltA11yHelper, so that CobaltA11yHelper can trigger `OnKey` function
on CobaltActivity.

We have to manually invoke imeAdapter.dispatchKeyEvent() for
KEYCODE_DPAD in class CobaltActivity.

video demo:
https://youtu.be/rzkOxCgk-_A?si=Y-pfS541gp0MEIgt

b/391708407