Run TTS cb with stub value for Starboard platforms (#5279)
The IsTextToSpeechEnabledSync Mojo method is not yet implemented for
Starboard platforms but we should at least run the callback with a stub
value. We are currently seeing a crash on the main branch because the
callback is getting destroyed without first being run or its binding
getting closed. Thanks joeltine for flagging.

This is a fix-forward from #5214.

b/391708407