Fixing unusable app after network error popup (#5513)
b/412443945

Addressing a bug where the app shows a blank screen and becomes unusable
after pressing the back button on the Network Error Dialog popup. Was
caused by undefined behavior upon dismissal of the popup; moved the
reloading logic from the RETRY_BUTTON to the onDismiss() function to
have the Navigation controller reload in both cases.