Create singleton class AudioOutputManager in native (#5810)
This is the first step to migrate AudioOutputManager to
jni_generator/jni_zero.

AudioOutputManager is a global singleton, I created the corresponding
AudioOutputManager class in native, which adheres to jni_zero coding
standard.
The main purpose of the native class AudioOutputManager is to manage the
Java object:
ScopedJavaGlobalRef<jobject> j_audio_output_manager_;, and encapsulate
all the @CallByNative functions.

I only migrated the inner class OutputDeviceInfo in AudioOutputManager
in this first PR, will update the rest functions to @CallByNative, and
update their invocation in follow-up PRs.

This PR does not modify how the AudioOutputManager class is invoked in
native code, and I verified the PR by playing a playback locally.

b/418059619