Disable V8 concurrent marking (#5805)
Marking TracedReference concurrently resulted in a significant crash in
Dogfood. There's no easy fix but we are confident the root cause is
concurrent marking for TracedReference.

Due to the significant volume of the crash, it might be better to
disable concurrent marking until either rebase or a fix becomes
available.

b/415843979