CSP Headers enforcement (#5879)
b/413043661

Enforcing CSP Headers during navigation in Chrobalt using the navigation
throttle implemented in #5766. Similarly makes the app show a blank
white screen in an unusable state when navigating to a URL that doesn't
have CSP headers in its response. Only on by default in gold builds, can 
use a command-line switch to enforce it as well.