Refactor to Split Parsing for Feature flags and CLI Switches (#5937)
This modifies the way Cobalt accepts command line args. Prior to this,
it would allow for multiple entries for any of those fields and was
impossible to set multiple `enable-features`, among others.

Since Cobalt has several different feature flags and command line
switches it hard-codes into the APK, and also accepts others from the
command line, this makes it so the CobaltActivity accumulates
`enable-features`, `disable-features`, `blink-enable-features`, and
`js-flags` separately using different command line fields.

This refactors the commandline overrides into a separate utility class,
making it more modular and testable. The junit tests for this are also
added.

Follow-ups include adding more tests to the class and modifying the
android monitoring script under //cobalt//tools/performance to support
the expected format of the various fields.

Test: `out/android-arm_devel/bin/run_cobalt_coat_junit_tests`
b/420947072