Add diff workflow to Cobalt GitHub workflows. (#5812)
The diff workflow performs a diff between a branch/PR and one or more
Chromium branches. This workflow triggers when pull requests are labeled
with diff-<milestone> where milestone is the name of the milestone (e.g.
'diff-m114'). Alternatively, the workflow can be dispatched manually by
providing a target branch name and comma separated list of milestones to
diff.

For now, this diff only generates diff files and attempts to print them
to stdout. In the future, we will export these files to a proper data
store for analysis.

b/409339952